+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 public final class Knight extends Piece {
