+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 public abstract class Piece {
