+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 public final class Queen extends Piece {
