+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,4 +1,4 @@
-package chess.piece;
+package chess.domain.piece;
 
 public final class Rook extends Piece {
