+++ src/test/java/chess/domain/board/PositionTest.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain.board;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
